Move command_queue_common
