+++ tests/tt_metal/tt_metal/device/test_device.cpp
@@ -20,7 +20,7 @@
 #include <vector>
 
 #include <tt-metalium/buffer_types.hpp>
-#include <tt-metalium/command_queue_common.hpp>
+#include "impl/dispatch/command_queue_common.hpp"
 #include <tt-metalium/core_coord.hpp>
 #include <tt-metalium/data_types.hpp>
 #include <tt-metalium/device.hpp>
