+++ tt_metal/CMakeLists.txt
@@ -45,7 +45,6 @@ target_sources(
             api/tt-metalium/circular_buffer_constants.h
             api/tt-metalium/circular_buffer_types.hpp
             api/tt-metalium/command_queue.hpp
-            api/tt-metalium/command_queue_common.hpp
             api/tt-metalium/command_queue_interface.hpp
             api/tt-metalium/constants.hpp
             api/tt-metalium/core_coord.hpp
