+++ tt_metal/api/tt-metalium/command_queue_interface.hpp
@@ -12,5 +12,4 @@
 #include <tt-metalium/buffer.hpp>
 #include <umd/device/tt_core_coordinates.h>
 
-#include <tt-metalium/command_queue_common.hpp>
 #include <tt-metalium/system_memory_cq_interface.hpp>
