+++ tt_metal/impl/dispatch/dispatch_mem_map.hpp
@@ -5,11 +5,11 @@
 #pragma once
 
 #include <stdint.h>
-#include <tt-metalium/command_queue_common.hpp>
 #include <utility>
 #include <vector>
 
 #include <umd/device/tt_core_coordinates.h>
+#include "command_queue_common.hpp"
 #include "dispatch_settings.hpp"
 
 namespace tt {
