+++ tt_metal/impl/dispatch/system_memory_cq_interface.cpp
@@ -2,10 +2,10 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include <tt-metalium/command_queue_common.hpp>
 #include <tt-metalium/system_memory_cq_interface.hpp>
 
 #include "assert.hpp"
+#include "command_queue_common.hpp"
 #include "impl/context/metal_context.hpp"
 #include "dispatch_settings.hpp"
