+++ tt_metal/impl/event/dispatch.cpp
@@ -10,11 +10,11 @@
 #include <vector>
 
 #include "assert.hpp"
-#include "command_queue_common.hpp"
 #include "core_coord.hpp"
 #include "device.hpp"
 #include "impl/context/metal_context.hpp"
 #include "dispatch/kernels/cq_commands.hpp"
+#include "dispatch/command_queue_common.hpp"
 #include "dispatch/dispatch_settings.hpp"
 #include "dispatch_core_common.hpp"
 #include "hal_types.hpp"
